requirements/local.txt
-r tests.txt

django-debug-toolbar==3.4.0
